refactor(project): reorganize project structure and add dedicated models package
- Created 'flight_prep_assistant/src/models/' directory to consolidate all data models
- Moved existing model definitions (aircraft, airfield, api, briefing, recommendation, request_data) into the new 'models' package
- Added '__init__.py' for package initialization
- Updated imports in 'main.py' and 'helpers.py' to reflect new structure
- Introduced 'enums.py' for shared enumerations and refactored related code
- Improved code organization and separation of concerns between helpers, models, and core logic
