+++ flight_prep_assistant/src/models/briefing.py
@@ -1,0 +1,43 @@
+from uuid import UUID, uuid4
+
+from pydantic import BaseModel, Field
+
+from .aircraft import AircraftModel
+from .airfield import AirfieldModel
+from .recommendation import RecommendationBaseModel
+
+
+class BriefingModel(BaseModel):
+    """
+    Represents a complete preflight briefing package for a pilot.
+
+    This model consolidates essential information about the aircraft,
+    departure and arrival airfields, and provides a computed recommendation
+    based on current operational and meteorological conditions.
+
+    Attributes:
+        briefing_id (UUID): A unique identifier for the generated briefing instance.
+        aircraft (AircraftModel): Detailed specifications and performance data for the selected aircraft.
+        departure_airfield (AirfieldModel): Information about the departure airfield,
+            including runway configuration, weather, and operational parameters.
+        arrival_airfield (AirfieldModel): Information about the arrival airfield,
+            including runway configuration, weather, and operational parameters.
+        recommendation (RecommendationBaseModel): Operational summary and final flight recommendation
+            produced by the performance calculator, integrating aircraft and weather data
+            to determine flight feasibility under VFR or IFR conditions.
+    """
+
+    briefing_id: UUID = Field(default_factory=uuid4, description="Unique briefing ID.")
+    aircraft: AircraftModel = Field(
+        description="Information about the aircraft used for the flight."
+    )
+    departure_airfield: AirfieldModel = Field(
+        description="Data about the departure airfield including weather."
+    )
+    arrival_airfield: AirfieldModel = Field(
+        description="Data about the arrival airfield including weather."
+    )
+    recommendation: str = Field(
+        description="Operational insights and the final flight recommendation generated by the "
+                    "performance calculator, summarizing key findings from the preflight briefing."
+    )
